# api-doc.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "utoipa-actix-web",
    "description": "Utoipa's actix-web bindings for seamless integration of the two",
    "contact": {
      "name": "Juha Kukkonen",
      "email": "juha7kukkonen@gmail.com"
    },
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.1.2"
  },
  "paths": {
...
    "/houses/{id}": {
      "get": {
        "operationId": "get_by_id",
        "responses": {
          "200": {
            "description": "House details",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "House not found"
          }
        }
      },
      "post": {
        "operationId": "compose",
        "responses": {
          "200": {
            "description": "House composed successfully"
          },
          "500": {
            "description": "Error composing house"
          }
        }
      }
    },
    "/houses/{id}/battery/{id}": {
      "get": {
        "operationId": "get_by_id",
        "parameters": [
          {
            "name": "house_id",
            "in": "path",
            "description": "House ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "battery_id",
            "in": "path",
            "description": "Battery ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get battery properties",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatteryInfo"
                }
              }
            }
          },
          "400": {
            "description": "Invalid battery ID"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "operationId": "add_by_id",
        "parameters": [
          {
            "name": "house_id",
            "in": "path",
            "description": "House ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "entity_name",
            "in": "path",
            "description": "Name of the battery entity",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatteryEntityParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Battery added successfully"
          },
          "500": {
            "description": "Error adding battery"
          }
        }
      },
      "delete": {
        "operationId": "remove_by_id",
        "parameters": [
          {
            "name": "house_id",
            "in": "path",
            "description": "House ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "entity_name",
            "in": "path",
            "description": "Name of the battery entity",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Battery removed successfully"
          },
          "500": {
            "description": "Error removing battery"
          }
        }
      }
    },
...
    "/houses/{id}/config": {
      "post": {
        "operationId": "set_config",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "House config set successfully"
          },
          "500": {
            "description": "Error setting house config"
          }
        }
      }
    },
    "/houses/{id}/entities": {
      "get": {
        "operationId": "list_entities",
        "responses": {
          "200": {
            "description": "List of entities",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Error getting entities"
          }
        }
      }
    },
...
    "/houses/{id}/houses/{id}": {
      "delete": {
        "operationId": "reset",
        "responses": {
          "200": {
            "description": "House reset successfully"
          },
          "500": {
            "description": "Error resetting house"
          }
        }
      }
    },
    "/houses/{id}/load": {
      "post": {
        "operationId": "load",
        "responses": {
          "200": {
            "description": "House loaded successfully"
          },
          "500": {
            "description": "Error loading house"
          }
        }
      }
    },
    "/houses/{id}/meters/{id}": {
      "get": {
        "operationId": "get_by_id",
        "responses": {
          "200": {
            "description": "Get meter information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeterInfo"
                }
              }
            }
          },
          "500": {
            "description": "Failed to get meter information"
          }
        }
      }
    },
...
    "/houses/{id}/pause": {
      "post": {
        "operationId": "pause_simulation",
        "responses": {
          "200": {
            "description": "House paused successfully"
          },
          "500": {
            "description": "Error pausing house"
          }
        }
      }
    },
    "/houses/{id}/resume": {
      "post": {
        "operationId": "resume_simulation",
        "responses": {
          "200": {
            "description": "House resumed successfully"
          },
          "500": {
            "description": "Error resuming house"
          }
        }
      }
    },
    "/houses/{id}/solar/{id}": {
      "get": {
        "operationId": "get_by_id",
        "responses": {
          "200": {
            "description": "Get solar information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolarInfo"
                }
              }
            }
          },
          "500": {
            "description": "Failed to get solar information"
          }
        }
      },
      "post": {
        "operationId": "add_by_id",
        "parameters": [
          {
            "name": "house_id",
            "in": "path",
            "description": "House ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "entity_name",
            "in": "path",
            "description": "Name of the solar entity",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolarEntityParams"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Add solar entity successfully"
          },
          "500": {
            "description": "Failed to add solar entity"
          }
        }
      },
      "delete": {
        "operationId": "remove_by_id",
        "parameters": [
          {
            "name": "house_id",
            "in": "path",
            "description": "House ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "entity_name",
            "in": "path",
            "description": "Name of the solar entity",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Remove solar entity successfully"
          },
          "500": {
            "description": "Failed to remove solar entity"
          }
        }
      }
    },
...
    "/houses/{id}/stop": {
      "post": {
        "operationId": "stop_simulation",
        "responses": {
          "200": {
            "description": "House stopped successfully"
          },
          "500": {
            "description": "Error stopping house"
          }
        }
      }
    },
    "/houses/{id}/thermal/{id}": {
      "get": {
        "operationId": "get_by_id",
        "parameters": [
          {
            "name": "house_id",
            "in": "path",
            "description": "House ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "thermal_id",
            "in": "path",
            "description": "Thermal ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get thermal information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThermalInfo"
                }
              }
            }
          },
          "500": {
            "description": "Failed to get thermal information"
          }
        }
      }
    },
...
    "/houses/{id}/time": {
      "get": {
        "operationId": "get_time",
        "responses": {
          "200": {
            "description": "Current time",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Error getting time"
          }
        }
      },
      "post": {
        "operationId": "set_time",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Time"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "House time set successfully"
          },
          "500": {
            "description": "Error setting house time"
          }
        }
      }
    },
    "/houses/{id}/timeshifters/{entity_name}": {
      "get": {
        "operationId": "get_by_id",
        "parameters": [
          {
            "name": "house_id",
            "in": "path",
            "description": "House ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "entity_name",
            "in": "path",
            "description": "Name of the timeshifter entity",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get timeshifter properties",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceStatus"
                }
              }
            }
          },
          "400": {
            "description": "Invalid entity name"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BatteryEntityParams": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "BatteryInfo": {
        "type": "object",
        "required": [
          "capacity",
          "max_charge",
          "max_discharge",
          "state_of_charge",
          "status",
          "consumption"
        ],
        "properties": {
          "capacity": {
            "type": "number",
            "format": "double"
          },
          "consumption": {
            "type": "number",
            "format": "double"
          },
          "max_charge": {
            "type": "number",
            "format": "double"
          },
          "max_discharge": {
            "type": "number",
            "format": "double"
          },
          "state_of_charge": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "$ref": "#/components/schemas/BatteryStatus"
          },
          "target_soc": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          }
        }
      },
      "BatteryStatus": {
        "type": "string",
        "enum": [
          "Charging",
          "Discharging",
          "Idle"
        ]
      },
      "DeviceStatus": {
        "type": "object",
        "required": [
          "house_id",
          "entity_name",
          "is_active",
          "progress",
          "active_job_idx",
          "scheduled_jobs",
          "consumption",
          "profile"
        ],
        "properties": {
          "active_job": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Job"
              }
            ]
          },
          "active_job_idx": {
            "type": "integer",
            "format": "int32"
          },
          "consumption": {
            "$ref": "#/components/schemas/Measurement"
          },
          "entity_name": {
            "type": "string"
          },
          "house_id": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "is_active": {
            "type": "boolean"
          },
          "profile": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InternalComplex"
            }
          },
          "progress": {
            "type": "number",
            "format": "double"
          },
          "scheduled_jobs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Job"
            }
          }
        }
      },
...
      "InternalComplex": {
        "type": "object",
        "required": [
          "re",
          "im"
        ],
        "properties": {
          "im": {
            "type": "number",
            "format": "double"
          },
          "re": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "Job": {
        "type": "object",
        "required": [
          "startTime",
          "endTime"
        ],
        "properties": {
          "endTime": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "startTime": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "Measurement": {
        "type": "object",
        "required": [
          "value",
          "unit"
        ],
        "properties": {
          "unit": {
            "type": "string"
          },
          "value": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "MeterInfo": {
        "type": "object",
        "required": [
          "house_id",
          "meter_id",
          "total_import",
          "total_export"
        ],
        "properties": {
          "current_export": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "current_import": {
            "type": [
              "number",
              "null"
            ],
            "format": "double"
          },
          "house_id": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "meter_id": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "total_export": {
            "type": "number",
            "format": "double"
          },
          "total_import": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "SimConfig": {
        "type": "object",
        "required": [
          "timeDelayBase",
          "timeBase",
          "timeOffset",
          "timeZone",
          "intervals",
          "startTime",
          "database",
          "dataPrefix",
          "clearDB",
          "extendedLogging",
          "logDevices",
          "logFlow",
          "enablePersistence",
          "weatherFile",
          "irradianceFile",
          "ventilationFile",
          "gainFile",
          "dhwFile",
          "houseNum",
          "useIslanding",
          "photoVoltaicSettings",
          "batterySettings",
          "heatingSettings",
          "useFillMethod",
          "usePP",
          "ctrlTimeBase",
          "thermostatStartTimes",
          "thermostatSetpoints"
        ],
        "properties": {
          "batterySettings": {
            "type": "string"
          },
          "clearDB": {
            "type": "boolean"
          },
          "ctrlTimeBase": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "dataPrefix": {
            "type": "string"
          },
          "database": {
            "type": "string"
          },
          "dhwFile": {
            "type": "string"
          },
          "enablePersistence": {
            "type": "boolean"
          },
          "extendedLogging": {
            "type": "boolean"
          },
          "gainFile": {
            "type": "string"
          },
          "heatingSettings": {
            "type": "string"
          },
          "houseNum": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "intervals": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "irradianceFile": {
            "type": "string"
          },
          "logDevices": {
            "type": "boolean"
          },
          "logFlow": {
            "type": "boolean"
          },
          "photoVoltaicSettings": {
            "type": "string"
          },
          "startTime": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "thermostatSetpoints": {
            "type": "string"
          },
          "thermostatStartTimes": {
            "type": "string"
          },
          "timeBase": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "timeDelayBase": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "timeOffset": {
            "type": "integer",
            "format": "int64"
          },
          "timeZone": {
            "type": "string"
          },
          "useFillMethod": {
            "type": "boolean"
          },
          "useIslanding": {
            "type": "boolean"
          },
          "usePP": {
            "type": "boolean"
          },
          "ventilationFile": {
            "type": "string"
          },
          "weatherFile": {
            "type": "string"
          }
        }
      },
      "SolarEntityParams": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "SolarInfo": {
        "type": "object",
        "required": [
          "consumption"
        ],
        "properties": {
          "consumption": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "ThermalInfo": {
        "type": "object",
        "required": [
          "current_temperature",
          "target_temperature",
          "heating_power",
          "consumption"
        ],
        "properties": {
          "consumption": {
            "type": "number",
            "format": "double"
          },
          "current_temperature": {
            "type": "number",
            "format": "double"
          },
          "heating_power": {
            "type": "number",
            "format": "double"
          },
          "target_temperature": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "Time": {
        "type": "object",
        "required": [
          "time"
        ],
        "properties": {
          "time": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
    }
  }
}